move new_transaction
