+++ src/new_transaction/tree/user_asset.rs
@@ -10,8 +10,10 @@ use crate::{
         sparse_merkle_tree::{MerklePath, Node},
         tree::MerkleProof,
     },
+    new_transaction::asset::{
+        encode_contributed_asset, ContributedAsset, TokenKind, VariableIndex,
+    },
     sparse_merkle_tree::{goldilocks_poseidon::le_bytes_to_bits, tree::KeyLike},
-    transaction::asset::{encode_contributed_asset, ContributedAsset, TokenKind, VariableIndex},
     zkdsa::account::Address,
 };
 
@@ -320,8 +322,8 @@ fn test_prove_user_asset_tree() {
 
     use crate::{
         merkle_tree::tree::get_merkle_root,
+        new_transaction::asset::{ContributedAsset, TokenKind, VariableIndex},
         sparse_merkle_tree::goldilocks_poseidon::GoldilocksHashOut,
-        transaction::asset::{ContributedAsset, TokenKind, VariableIndex},
         zkdsa::account::{private_key_to_account, Address},
     };
 
@@ -395,11 +397,11 @@ fn test_user_asset_tree_by_plonky2() {
 
     use crate::{
         merkle_tree::tree::get_merkle_root,
-        sparse_merkle_tree::goldilocks_poseidon::GoldilocksHashOut,
-        transaction::{
+        new_transaction::{
             asset::{ContributedAsset, TokenKind, VariableIndex},
             tree::user_asset::UserAssetTree,
         },
+        sparse_merkle_tree::goldilocks_poseidon::GoldilocksHashOut,
         zkdsa::account::{private_key_to_account, Address},
     };
